requirements.txt
# =============================================================================
# Khabri - News Intelligence System
# Python Dependencies
# =============================================================================

# Core
python-dotenv>=1.0.0
PyYAML>=6.0
pytz>=2023.3

# Async (used in scripts)
aiohttp>=3.9.0

# Web Scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0
feedparser>=6.0.0
requests>=2.31.0

# Email (built-in smtplib is used, no external SMTP lib needed)

# Utilities
python-dateutil>=2.8.0

# =============================================================================
# Optional Features (recommended for full functionality)
# =============================================================================
# Google Trends integration for keyword research
# If not installed, keyword engine falls back to heuristic extraction
pytrends>=4.9.0

# =============================================================================
# Development & Testing (optional)
# =============================================================================
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# pytest-cov>=4.1.0
# black>=23.0.0
# flake8>=6.0.0
